B-1 refactoring path and packages.
